# services/app_api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdjustedOut": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/AdjustedRowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "rows"
        ],
        "title": "AdjustedOut",
        "type": "object"
      },
      "AdjustedRowOut": {
        "properties": {
          "adjusted_date": {
            "format": "date",
            "title": "Adjusted Date",
            "type": "string"
          },
          "batch_no": {
            "title": "Batch No",
            "type": "string"
          },
          "delta_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Days"
          },
          "material_desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Desc"
          },
          "material_no": {
            "title": "Material No",
            "type": "string"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "reason_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Label"
          },
          "row_key": {
            "title": "Row Key",
            "type": "string"
          },
          "set_at": {
            "format": "date-time",
            "title": "Set At",
            "type": "string"
          },
          "set_by": {
            "title": "Set By",
            "type": "string"
          },
          "system_need_by_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Need By Date"
          }
        },
        "required": [
          "row_key",
          "batch_no",
          "material_no",
          "material_desc",
          "system_need_by_date",
          "adjusted_date",
          "delta_days",
          "reason_code",
          "reason_label",
          "set_by",
          "set_at"
        ],
        "title": "AdjustedRowOut",
        "type": "object"
      },
      "AlertOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "detail": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Detail",
            "type": "object"
          },
          "kind": {
            "enum": [
              "air_gap",
              "late",
              "on_hold",
              "rejected"
            ],
            "title": "Kind",
            "type": "string"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RowOut"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "count"
        ],
        "title": "AlertOut",
        "type": "object"
      },
...
      "FlowEntry": {
        "properties": {
          "breached": {
            "title": "Breached",
            "type": "boolean"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "late_count": {
            "title": "Late Count",
            "type": "integer"
          },
          "skip_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Count"
          },
          "stage_key": {
            "title": "Stage Key",
            "type": "string"
          }
        },
        "required": [
          "stage_key",
          "label",
          "count",
          "breached",
          "late_count"
        ],
        "title": "FlowEntry",
        "type": "object"
      },
...
      "Freshness": {
        "description": "Which contract run the answer is built from, and how old it is on the demo clock (F09-FR-07).",
        "properties": {
          "contract_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Run Id"
          },
          "freshness_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness Minutes"
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          }
        },
        "required": [
          "contract_run_id",
          "last_success_at",
          "freshness_minutes"
        ],
        "title": "Freshness",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InsightRowOut": {
        "properties": {
          "air_gap_hours": {
            "title": "Air Gap Hours",
            "type": "integer"
          },
          "batch_no": {
            "title": "Batch No",
            "type": "string"
          },
          "days_gap": {
            "title": "Days Gap",
            "type": "integer"
          },
          "material_desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Desc"
          },
          "material_no": {
            "title": "Material No",
            "type": "string"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "row_key": {
            "title": "Row Key",
            "type": "string"
          },
          "stage_key": {
            "title": "Stage Key",
            "type": "string"
          },
          "stage_label": {
            "title": "Stage Label",
            "type": "string"
          }
        },
        "required": [
          "row_key",
          "batch_no",
          "material_no",
          "material_desc",
          "stage_key",
          "stage_label",
          "air_gap_hours",
          "days_gap"
        ],
        "title": "InsightRowOut",
        "type": "object"
      },
      "InsightsOut": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/InsightRowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "rows"
        ],
        "title": "InsightsOut",
        "type": "object"
      },
...
      "OverviewOut": {
        "properties": {
          "adjusted_count": {
            "title": "Adjusted Count",
            "type": "integer"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Alerts",
            "type": "array"
          },
          "batch_count": {
            "title": "Batch Count",
            "type": "integer"
          },
          "bookmarks": {
            "items": {
              "type": "string"
            },
            "title": "Bookmarks",
            "type": "array"
          },
          "flow_strip": {
            "items": {
              "$ref": "#/components/schemas/FlowEntry"
            },
            "title": "Flow Strip",
            "type": "array"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "mode": {
            "enum": [
              "snapshot",
              "due_in_period"
            ],
            "title": "Mode",
            "type": "string"
          },
          "on_hold_count": {
            "title": "On Hold Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "freshness",
          "flow_strip",
          "on_hold_count",
          "adjusted_count",
          "total",
          "batch_count",
          "mode",
          "alerts",
          "bookmarks",
          "rows"
        ],
        "title": "OverviewOut",
        "type": "object"
      },
...
      "ProposalRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "priority"
        ],
        "title": "ProposalRef",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "R2R Intelligence application API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/overview": {
      "get": {
        "operationId": "overview_api_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "type[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type[]"
            }
          },
          {
            "in": "query",
            "name": "class[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class[]"
            }
          },
          {
            "in": "query",
            "name": "campaign[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campaign[]"
            }
          },
          {
            "in": "query",
            "name": "flags[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flags[]"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "in": "query",
            "name": "include_released",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Released",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "bookmarked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Bookmarked",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview"
      }
    },
    "/api/overview/adjusted": {
      "get": {
        "description": "The Adjusted Needs-by Dates window (F16-FR-07): every filter except stage applies.",
        "operationId": "adjusted_api_overview_adjusted_get",
        "parameters": [
          {
            "in": "query",
            "name": "type[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type[]"
            }
          },
          {
            "in": "query",
            "name": "class[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class[]"
            }
          },
          {
            "in": "query",
            "name": "campaign[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campaign[]"
            }
          },
          {
            "in": "query",
            "name": "flags[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flags[]"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "in": "query",
            "name": "include_released",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Released",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "bookmarked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Bookmarked",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adjusted"
      }
    },
    "/api/overview/insights": {
      "get": {
        "description": "The air-gap insights window (F16-FR-09): every air-gap row, worst first, no cap.",
        "operationId": "insights_api_overview_insights_get",
        "parameters": [
          {
            "in": "query",
            "name": "type[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type[]"
            }
          },
          {
            "in": "query",
            "name": "class[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class[]"
            }
          },
          {
            "in": "query",
            "name": "campaign[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campaign[]"
            }
          },
          {
            "in": "query",
            "name": "flags[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flags[]"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "in": "query",
            "name": "include_released",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Released",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "bookmarked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Bookmarked",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Insights"
      }
    },
...
  }
}